% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_datatable.R
\name{as_datatable}
\alias{as_datatable}
\title{Convert a list of \code{xts} objects into a single \code{data.table} for CrowdQC+}
\usage{
as_datatable(x = NULL)
}
\arguments{
\item{x}{List of \code{xts} objects as provided by \code{get_measure()}.}
}
\value{
Data.table object.
}
\description{
Convert a list of \code{xts} objects into a single \code{data.table} for CrowdQC+
}
\examples{
\dontrun{
dt <- as_datatable(meas)
}
}
