% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_token.R
\name{fetch_token}
\alias{fetch_token}
\title{Fetch an Oauth 2.0 token from api.netatmo.com}
\usage{
fetch_token()
}
\value{
A \code{Token2.0} object.
}
\description{
Fetch an Oauth 2.0 token from api.netatmo.com
}
\examples{
\dontrun{
fetch_token()
}
}
