% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_extent}
\alias{get_extent}
\title{Construct an object of type bbox using coordinates or specific polygons}
\usage{
get_extent(input, crs = 4326)
}
\arguments{
\item{input}{Vector of length 4 containing numeric representing coordinates,
string of length 1 representing the name of a municipality,
or string of length 5 representing a postal zip code.}

\item{crs}{(optional) Coordinate reference system definition.}
}
\value{
An object of type \code{bbox}.
}
\description{
Construct an object of type bbox using coordinates or specific polygons
}
\examples{
\dontrun{
bbox1 <- get_extent(input = c(6.89, 51.34, 7.13, 51.53), crs = 4326)
bbox2 <- get_extent(input = "Essen")
bbox3 <- get_extent(input = "45145")
}
}
