% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure.R
\name{get_measure}
\alias{get_measure}
\title{Title}
\usage{
get_measure(devices, parameter = NULL, resolution = 5, period = NULL)
}
\arguments{
\item{devices}{sf object}

\item{parameter}{character}

\item{period}{date}
}
\value{
xts
}
\description{
Title
}
\examples{
get_measure(stations, parameter = "pressure")
get_measure(stations, parameter = "temperature", resolution = 30, period = "recent")
get_measure(stations, parameter = "sum_rain", resolution = 60, period = c("2022-06-06", "2022-06-08"))
}
