% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_n_queries.R
\name{get_n_queries}
\alias{get_n_queries}
\title{Determine the number of individual requests needed to cover defined specifications}
\usage{
get_n_queries(d = NULL, res = NULL, p = NULL)
}
\arguments{
\item{d}{numeric. Number of stations.}

\item{res}{numeric. Measurement resolution in minutes.}

\item{p}{numeric. Vector of length 2 representing an interval definition in UNIX time.}
}
\value{
numeric. Number of necessary requests.
}
\description{
Determine the number of individual requests needed to cover defined specifications
}
\examples{
period <- get_period("2024-02-01/2024-04-01")

# 10 stations, hourly data
get_n_queries(d = 10, res = 60, p = period)

# 20 stations, 5-minutely data
get_n_queries(d = 20, res = 5, p = period)
}
