% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2_token.R
\name{get_oauth2_token}
\alias{get_oauth2_token}
\title{Title}
\usage{
get_oauth2_token(file)
}
\arguments{
\item{file}{}
}
\value{
Auth token: Token2.0
}
\description{
Title
}
\examples{
get_oauth2_token("oauth.cfg")
}
