% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2token.R
\name{get_oauth2token}
\alias{get_oauth2token}
\title{Create an Oauth 2.0 token for api.netatmo.net}
\usage{
get_oauth2token(file)
}
\arguments{
\item{file}{Full path to oauth configuration file.}
}
\value{
A request object containing the Oauth 2.0 token.
}
\description{
Create an Oauth 2.0 token for api.netatmo.net
}
\examples{
get_oauth2token("oauth.cfg")
}
