% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_public_data.R
\name{get_public_data}
\alias{get_public_data}
\title{Title}
\usage{
get_public_data(extent, use_tiles = FALSE)
}
\arguments{
\item{use_tiles}{}
}
\value{
tibble
}
\description{
Title
}
\examples{
bbox <- get_extent(input = c(6.89, 51.34, 7.13, 51.53))
bbox <- get_extent(input = "Essen")
stations <- get_public_data(extent = bbox)
stations <- get_public_data(extent = bbox, use_tiles = TRUE)
}
