% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publicdata.R
\name{get_publicdata}
\alias{get_publicdata}
\title{Get Netatmo station locations with various metadata based on extent queried}
\usage{
get_publicdata(ext, use_tiles = FALSE)
}
\arguments{
\item{ext}{A bounding box as provided by \code{get_extent()}.}

\item{use_tiles}{Boolean. Should fetching be done in spatial slices? More results are to be expected when \code{TRUE}.}
}
\value{
A simple feature collection containing station metadata and geometries.
}
\description{
Get Netatmo station locations with various metadata based on extent queried
}
\examples{
e <- get_extent(x = c(6.89, 51.34, 7.13, 51.53))
stations <- get_publicdata(ext = e)
stations <- get_publicdata(ext = e, use_tiles = TRUE)
}
