% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2token.R
\name{is_expired}
\alias{is_expired}
\title{Check if your Oauth 2.0 token is expired and needs to be refreshed}
\usage{
is_expired()
}
\value{
A boolean.
}
\description{
Check if your Oauth 2.0 token is expired and needs to be refreshed
}
\examples{
is_expired()
}
\keyword{internal}
