% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2token.R
\name{print_at}
\alias{print_at}
\title{Return access token to console for further use in a browser}
\usage{
print_at()
}
\value{
character. Access token representation.
}
\description{
Return access token to console for further use in a browser
}
\examples{
\dontrun{
print_at()
}
}
