% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2token.R
\name{print_rt}
\alias{print_rt}
\title{Return your refresh token as a string for further use in a browser}
\usage{
print_rt()
}
\value{
A string representing the refresh token.
}
\description{
Return your refresh token as a string for further use in a browser
}
\examples{
print_rt()
}
