% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2token.R
\name{refresh_at}
\alias{refresh_at}
\title{Refresh your access token using the refresh token}
\usage{
refresh_at()
}
\value{
The refreshed token.
}
\description{
Refresh your access token using the refresh token
}
\examples{
refresh_at()
}
\keyword{internal}
