% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Store Netatmo API credentials (id/secret) using \code{{keyring}}}
\usage{
set_credentials(update = FALSE)
}
\arguments{
\item{update}{logical. Enforce setting new credentials despite of existing?}
}
\description{
Store Netatmo API credentials (id/secret) using \code{{keyring}}
}
\examples{
\dontrun{
set_credentials()
}
}
\seealso{
\code{\link{keyring}}
}
