% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_device.R
\name{set_device}
\alias{set_device}
\title{Manually construct an sf object with user-provided MAC addresses}
\usage{
set_device(
  base_station = NULL,
  NAModule1 = NULL,
  NAModule2 = NULL,
  NAModule3 = NULL,
  lat = NULL,
  lon = NULL
)
}
\arguments{
\item{base_station}{character. MAC address of Netatmo base station.}

\item{NAModule1}{character. MAC address of Netatmo outdoor module.}

\item{NAModule2}{character. MAC address of Netatmo wind module.}

\item{NAModule3}{character. MAC address of Netatmo precipitation module.}

\item{lat}{(optional) numeric. Geodetic latitude in EPSG: 4326.}

\item{lon}{(optional) numeric. Geodetic longitude in EPSG: 4326.}
}
\value{
Sf object containing relevant MAC addresses for \code{get_measure()} queries.
}
\description{
Manually construct an sf object with user-provided MAC addresses
}
\examples{
set_device("70:ee:50:13:54:bc")
set_device("70:ee:50:13:54:bc", "02:00:00:13:57:c8", "06:00:00:02:5f:54", "05:00:00:01:48:96")
}
