% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vg250_gem}
\alias{vg250_gem}
\title{Polygon representation of municipalities in Germany (VG250 31.12.2021)}
\format{
Simple feature collection of type MULTIPOLYGON with 10,994 features and 1 field:
\describe{
\item{GEN}{municipality name}
\item{geometry}{centroid coordinates}
}
}
\source{
\url{https://gdz.bkg.bund.de/index.php/default/digitale-geodaten/verwaltungsgebiete/verwaltungsgebiete-1-250-000-stand-31-12-vg250-31-12.html}
}
\usage{
vg250_gem
}
\description{
A subset of data from the VG250_GEM product provided by the Federal Agency for Cartography and Geodesy, Germany
}
\details{
: pk <- sf::st_read("VG250_PK.shp")

: vg250_gem <- sf::st_read("VG250_GEM.shp") |> sf::st_filter(x = _, y = pk) |> dplyr::select("GEN") |> sf::st_transform("epsg:4326")
}
\note{
: License: Data licence Germany – attribution – version 2.0

: Copyright: GeoBasis-DE / BKG 2022 (modified)
}
\keyword{datasets}
