% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_token.R
\name{fetch_token}
\alias{fetch_token}
\title{Fetch an Oauth 2.0 token from api.netatmo.net}
\usage{
fetch_token()
}
\value{
Request object containing the Oauth 2.0 token.
}
\description{
Fetch an Oauth 2.0 token from api.netatmo.net
}
\examples{
\dontrun{
fetch_token()
}
}
