% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_extent.R
\name{get_extent}
\alias{get_extent}
\title{Construct an object of type bbox based on user input}
\usage{
get_extent(x = NULL, epsg = 4326)
}
\arguments{
\item{x}{Vector of length 4 containing numeric representing coordinates (xmin, ymin, xmax, ymax),
or string of length 1 representing the name of a municipality,
or string of nchar 5 representing a postal zip code.}

\item{epsg}{(optional) Coordinate reference system definition.}
}
\value{
Object of type \code{sfc_POLYGON}.
}
\description{
Construct an object of type bbox based on user input
}
\examples{
e1 <- get_extent(x = c(6.89, 51.34, 7.13, 51.53))
e2 <- get_extent(x = c(353034.1, 5689295.3, 370288.6, 5710875.9), epsg = 25832)
e3 <- get_extent(x = "Essen")
e4 <- get_extent(x = "45145")
}
