% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure.R
\name{get_measure}
\alias{get_measure}
\title{Title}
\usage{
get_measure(
  devices,
  parameter = "sum_rain",
  resolution = 5,
  from = NULL,
  to = NULL
)
}
\arguments{
\item{devices}{sf object}

\item{parameter}{character}

\item{from}{date}

\item{to}{date}
}
\value{
xts
}
\description{
Title
}
\examples{
get_measure(stations, parameter = "pressure")
get_measure(stations, parameter = "temperature", resolution = 5)
get_measure(stations, parameter = "sum_rain", resolution = 5, from = "2022-04-04", to = "2022-04-06")
}
