% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure.R
\name{get_measure}
\alias{get_measure}
\title{Retrieve data from a device or module}
\usage{
get_measure(devices = NULL, period = NULL, par = NULL, res = 5)
}
\arguments{
\item{devices}{Sf object as provided by \code{get_publicdata()}.}

\item{period}{numeric. From/to period vector as provided by \code{get_period()}.}

\item{par}{character. Meteorological parameter to query.}

\item{res}{numeric. Measurement resolution in minutes.}
}
\value{
List of \code{xts} objects.
}
\description{
Retrieve data from a device or module
}
\examples{
\dontrun{
fetch_token()

e <- get_extent(x = c(6.89, 51.34, 7.13, 51.53))
stations <- get_publicdata(ext = e)

p1 <- get_period()
p2 <- get_period(res = 60)
p3 <- get_period(x = "recent")
p4 <- get_period(x = c("2022-06-06", "2022-06-08"))

meas <- get_measure(stations, period = p2, par = "pressure")
meas <- get_measure(stations, period = p3, par = "temperature", res = 30)
meas <- get_measure(stations, period = p4, par = "sum_rain", res = 60)
}
}
