% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_measure.R
\name{get_measure}
\alias{get_measure}
\title{Get Netatmo station observations}
\usage{
get_measure(devices, par = NULL, res = 5, period = NULL)
}
\arguments{
\item{devices}{A \code{sf} object as provided by \code{get_publicdata()}.}

\item{par}{The meteorological parameter to acquire.}

\item{res}{Measurement resolution in minutes.}

\item{period}{A from/to period vector as provided by \code{get_period()}.}
}
\value{
A list of \code{xts} objects.
}
\description{
Get Netatmo station observations
}
\examples{
get_measure(stations, par = "pressure")
get_measure(stations, par = "temperature", res = 30, period = "recent")
get_measure(stations, par = "sum_rain", res = 60, period = c("2022-06-06", "2022-06-08"))
}
