% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_period}
\alias{get_period}
\title{Title}
\usage{
get_period(input = NULL, resolution = 5)
}
\arguments{
\item{resolution}{}
}
\description{
Title
}
\examples{
get_period()
get_period(input = "recent")
get_period(input = c("2022-06-01", "2022-06-04"))
}
