% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_period.R
\name{get_period}
\alias{get_period}
\title{Construct a vector of length 2 and integer type representing UNIX time}
\usage{
get_period(x = NULL, res = 5)
}
\arguments{
\item{x}{NULL, or "recent", or a character vector of length 2 representing two timestamps (from/to).}

\item{res}{numeric. Measurement resolution in minutes.}
}
\value{
numeric. Vector of length 2 containing from/to timestamps as UNIX time.
}
\description{
Construct a vector of length 2 and integer type representing UNIX time
}
\examples{
p1 <- get_period()
p2 <- get_period(res = 60)
p3 <- get_period("recent")
p4 <- get_period(c("2024-03-01", "2024-04-01"))
}
