% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_period}
\alias{get_period}
\title{Construct a vector of length 2 and integer type representing UNIX time}
\usage{
get_period(x = NULL, res = 5)
}
\arguments{
\item{x}{NULL, or "recent", or a vector of length 2 containing from/to timestamps as characters.}

\item{res}{numeric. Measurement resolution in minutes.}
}
\value{
numeric. Vector of length 2 containing from/to timestamps as UNIX time.
}
\description{
Construct a vector of length 2 and integer type representing UNIX time
}
\examples{
p1 <- get_period()
p2 <- get_period(x = "recent")
p3 <- get_period(x = c("2022-06-01", "2022-06-04"))
}
