% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_public_data.R
\name{get_public_data}
\alias{get_public_data}
\title{Title}
\usage{
get_public_data(bbox, use_tiles = FALSE)
}
\arguments{
\item{use_tiles}{}
}
\value{
tibble
}
\description{
Title
}
\examples{
stations <- get_public_data(bbox = "Essen", use_tiles = FALSE)
stations <- get_public_data(bbox = "Essen", use_tiles = TRUE)
}
