% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publicdata.R
\name{get_publicdata}
\alias{get_publicdata}
\title{Retrieves publicly shared weather data from outdoor modules within a predefined area}
\usage{
get_publicdata(ext = NULL, tiles = FALSE, cellsize = 0.05, meas = FALSE)
}
\arguments{
\item{ext}{Object of type \code{sfc_POLYGON}, as provided by \code{get_extent()}.}

\item{tiles}{logical. Should fetching be done in separate chunks? More results are to be expected using \code{TRUE}.}

\item{cellsize}{numeric. Edge length \code{[°]} of individual tiles to be used when \code{tiles = TRUE}.}

\item{meas}{logical. Include measurements returned by the API per default?}
}
\value{
Sf object containing station metadata.
}
\description{
Retrieves publicly shared weather data from outdoor modules within a predefined area
}
\examples{
\dontrun{
fetch_token()

e <- get_extent(x = c(6.89, 51.34, 7.13, 51.53))

stations <- get_publicdata(ext = e)
stations_tiles <- get_publicdata(ext = e, tiles = TRUE)
stations_meas <- get_publicdata(ext = e, meas = TRUE)
}
}
\seealso{
\code{\link[=get_extent]{get_extent()}}
}
