% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_publicdata.R
\name{get_publicdata}
\alias{get_publicdata}
\title{Get Netatmo station locations enriched with metadata}
\usage{
get_publicdata(ext, use_tiles = FALSE)
}
\arguments{
\item{ext}{Object of type \code{sfc_POLYGON}, as provided by \code{get_extent()}.}

\item{use_tiles}{logical. Should fetching be done in spatial slices? More results are to be expected using \code{TRUE}.}
}
\value{
Sf object containing station metadata and geometries.
}
\description{
Get Netatmo station locations enriched with metadata
}
\examples{
get_oauth2token("oauth.cfg")

e <- get_extent(x = c(6.89, 51.34, 7.13, 51.53))

stations <- get_publicdata(ext = e)
stations <- get_publicdata(ext = e, use_tiles = TRUE)
}
