% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oauth2token.R
\name{print_at}
\alias{print_at}
\title{Return your access token as a string for further use in a browser}
\usage{
print_at()
}
\value{
A string representing the access token.
}
\description{
Return your access token as a string for further use in a browser
}
\examples{
print_at()
}
