% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_token.R
\name{print_rt}
\alias{print_rt}
\title{Return refresh token to console for further use in a browser}
\usage{
print_rt()
}
\value{
character. Refresh token representation.
}
\description{
Return refresh token to console for further use in a browser
}
\examples{
\dontrun{
print_rt()
}
}
