% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Store Netatmo API credentials (name/id/secret) using a \code{keyring}}
\usage{
set_credentials()
}
\value{
environment. creates a new keyring called \code{"netatmo"} with \code{"name"}, \code{"id"}, and \code{"secret"} stored.
}
\description{
Store Netatmo API credentials (name/id/secret) using a \code{keyring}
}
\examples{
set_credentials()
}
\seealso{
\code{\link{keyring}}
}
